backend/requirements.txt
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
uuid==1.30
Werkzeug==2.0.3
